More frequency peak to tone code.

+++ tests/frog_tones_peak.cc
@@ -127,7 +127,7 @@ TEST_F(FrogTonesPeakTest, HaveResultInPeakArray)
 {
 	freq_buffer[50] = 100;
 	ft->find_peaks();
-	EXPECT_EQ(peaks[0].freq, ft->bin_frequency(50));
+	EXPECT_EQ(peaks[0].freq, ft->bin_frequency((uint16_t)50));
 }
 
 TEST_F(FrogTonesPeakTest, HaveMidBinResultInPeakArray)
@@ -135,8 +135,8 @@ TEST_F(FrogTonesPeakTest, HaveMidBinResultInPeakArray)
 	freq_buffer[50] = 100;
 	freq_buffer[51] = 20;
 	ft->find_peaks();
-	EXPECT_GT((float)peaks[0].freq, (float)ft->bin_frequency(50));
-	EXPECT_LT((float)peaks[0].freq, (float)ft->bin_frequency(51));
+	EXPECT_GT((float)peaks[0].freq, (float)ft->bin_frequency((uint16_t)50));
+	EXPECT_LT((float)peaks[0].freq, (float)ft->bin_frequency((uint16_t)51));
 }
